+++ src/main/java/uk/gov/hmcts/reform/iacasenotificationsapi/infrastructure/clients/GovNotifyNotificationSender.java
@@ -2,16 +2,11 @@ package uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients;
 
 import static org.slf4j.LoggerFactory.getLogger;
 
-import com.github.benmanes.caffeine.cache.Cache;
-import com.github.benmanes.caffeine.cache.Caffeine;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.NotificationSender;
-import uk.gov.service.notify.NotificationClientException;
-import uk.gov.service.notify.SendEmailResponse;
-import uk.gov.service.notify.SendSmsResponse;
+import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients.helper.NotificationSenderHelper;
 
 @Service
 public class GovNotifyNotificationSender implements NotificationSender {
@@ -20,15 +15,16 @@ public class GovNotifyNotificationSender implements NotificationSender {
 
     private final int deduplicateSendsWithinSeconds;
     private final RetryableNotificationClient notificationClient;
-
-    private Cache<String, String> recentDeliveryReceiptCache;
+    private final NotificationSenderHelper senderHelper;
 
     public GovNotifyNotificationSender(
         @Value("${notificationSender.deduplicateSendsWithinSeconds}") int deduplicateSendsWithinSeconds,
-        RetryableNotificationClient notificationClient
+        RetryableNotificationClient notificationClient,
+        NotificationSenderHelper senderHelper
     ) {
         this.deduplicateSendsWithinSeconds = deduplicateSendsWithinSeconds;
         this.notificationClient = notificationClient;
+        this.senderHelper = senderHelper;
     }
 
     public synchronized String sendEmail(
@@ -37,40 +33,14 @@ public class GovNotifyNotificationSender implements NotificationSender {
         Map<String, String> personalisation,
         String reference
     ) {
-        recentDeliveryReceiptCache = getOrCreateDeliveryReceiptCache();
-        return recentDeliveryReceiptCache.get(
-            emailAddress + reference,
-            k -> {
-
-                try {
-
-                    LOG.info("Attempting to send email notification to GovNotify: {}", reference);
-
-                    SendEmailResponse response =
-                            notificationClient.sendEmail(
-                                    templateId,
-                                    emailAddress,
-                                    personalisation,
-                                    reference
-                            );
-
-                    String notificationId =
-                        response
-                            .getNotificationId()
-                            .toString();
-
-                    LOG.info(
-                        "Successfully sent email notification to GovNotify: {} ({})",
-                        reference,
-                        notificationId
-                    );
-
-                    return notificationId;
-
-                } catch (NotificationClientException e) {
-                    throw new NotificationServiceResponseException("Failed to send email using GovNotify", e);
-                }
-            }
+        return senderHelper.sendEmail(
+                templateId,
+                emailAddress,
+                personalisation,
+                reference,
+                notificationClient,
+                deduplicateSendsWithinSeconds,
+                LOG
         );
     }
 
@@ -80,54 +50,15 @@ public class GovNotifyNotificationSender implements NotificationSender {
         final String phoneNumber,
         final Map<String, String> personalisation,
         final String reference) {
-        recentDeliveryReceiptCache = getOrCreateDeliveryReceiptCache();
-        return recentDeliveryReceiptCache.get(
-            phoneNumber + reference,
-            k -> {
 
-                try {
-
-                    LOG.info("Attempting to send a text message notification to GovNotify: {}", reference);
-
-                    SendSmsResponse response =
-                            notificationClient
-                                    .sendSms(
-                                            templateId,
-                                            phoneNumber,
-                                            personalisation,
-                                            reference
-                                    );
-
-                    String notificationId =
-                        response
-                            .getNotificationId()
-                            .toString();
-
-                    LOG.info(
-                        "Successfully sent sms notification to GovNotify: {} ({})",
-                        reference,
-                        notificationId
-                    );
-
-                    return notificationId;
-
-                } catch (NotificationClientException e) {
-                    throw new NotificationServiceResponseException("Failed to send sms using GovNotify", e);
-                }
-            }
+        return senderHelper.sendSms(
+                templateId,
+                phoneNumber,
+                personalisation,
+                reference,
+                notificationClient,
+                deduplicateSendsWithinSeconds,
+                LOG
         );
-    }
-
-    private Cache<String, String> getOrCreateDeliveryReceiptCache() {
-
-        if (recentDeliveryReceiptCache == null) {
-            recentDeliveryReceiptCache =
-                Caffeine
-                    .newBuilder()
-                    .expireAfterWrite(deduplicateSendsWithinSeconds, TimeUnit.SECONDS)
-                    .build();
-        }
-
-        return recentDeliveryReceiptCache;
     }
 }
