+++ src/test/java/uk/gov/hmcts/reform/iacasenotificationsapi/infrastructure/clients/GovNotifyNotificationSenderTest.java
@@ -1,35 +1,33 @@
 package uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.awaitility.Awaitility.await;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static org.slf4j.LoggerFactory.getLogger;
 
 import java.util.Map;
 import java.util.UUID;
-import java.util.concurrent.TimeUnit;
-import org.awaitility.core.ConditionTimeoutException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients.helper.NotificationSenderHelper;
 import uk.gov.service.notify.NotificationClientException;
-import uk.gov.service.notify.SendEmailResponse;
-import uk.gov.service.notify.SendSmsResponse;
 
 @ExtendWith(MockitoExtension.class)
 @SuppressWarnings("unchecked")
 public class GovNotifyNotificationSenderTest {
 
+    private static final org.slf4j.Logger LOG = getLogger(GovNotifyNotificationSender.class);
+
     private int deduplicateSendsWithinSeconds = 1;
     @Mock
     private RetryableNotificationClient notificationClient;
+
+    @Mock private NotificationSenderHelper senderHelper;
 
     private String templateId = "a-b-c-d-e-f";
     private String emailAddress = "recipient@example.com";
@@ -44,7 +42,8 @@ public class GovNotifyNotificationSenderTest {
         govNotifyNotificationSender =
             new GovNotifyNotificationSender(
                 deduplicateSendsWithinSeconds,
-                notificationClient
+                notificationClient,
+                senderHelper
             );
     }
 
@@ -53,16 +52,15 @@ public class GovNotifyNotificationSenderTest {
 
         final UUID expectedNotificationId = UUID.randomUUID();
 
-        SendEmailResponse sendEmailResponse = mock(SendEmailResponse.class);
-
-        when(notificationClient.sendEmail(
-            templateId,
-            emailAddress,
-            personalisation,
-            reference
-        )).thenReturn(sendEmailResponse);
-
-        when(sendEmailResponse.getNotificationId()).thenReturn(expectedNotificationId);
+        when(senderHelper.sendEmail(
+                templateId,
+                emailAddress,
+                personalisation,
+                reference,
+                notificationClient,
+                deduplicateSendsWithinSeconds,
+                LOG
+        )).thenReturn(String.valueOf(expectedNotificationId));
 
         String actualNotificationId =
             govNotifyNotificationSender.sendEmail(
@@ -72,129 +70,17 @@ public class GovNotifyNotificationSenderTest {
                 reference
             );
 
-        verify(notificationClient, times(1)).sendEmail(
+        verify(senderHelper, times(1)).sendEmail(
                 templateId,
                 emailAddress,
                 personalisation,
-                reference
+                reference,
+                notificationClient,
+                deduplicateSendsWithinSeconds,
+                LOG
         );
 
         assertEquals(expectedNotificationId.toString(), actualNotificationId);
-    }
-
-    @Test
-    public void should_not_send_duplicate_emails_in_short_space_of_time() throws NotificationClientException {
-
-        final String otherEmailAddress = "foo@bar.com";
-        final String otherReference = "1111_SOME_OTHER_NOTIFICATION";
-
-        final UUID expectedNotificationId = UUID.randomUUID();
-        final UUID expectedNotificationIdForOther = UUID.randomUUID();
-
-        SendEmailResponse sendEmailResponse = mock(SendEmailResponse.class);
-        SendEmailResponse sendEmailResponseForOther = mock(SendEmailResponse.class);
-
-        when(notificationClient.sendEmail(
-            templateId,
-            emailAddress,
-            personalisation,
-            reference
-        )).thenReturn(sendEmailResponse);
-
-        when(notificationClient.sendEmail(
-            templateId,
-            otherEmailAddress,
-            personalisation,
-            otherReference
-        )).thenReturn(sendEmailResponseForOther);
-
-        when(sendEmailResponse.getNotificationId()).thenReturn(expectedNotificationId);
-        when(sendEmailResponseForOther.getNotificationId()).thenReturn(expectedNotificationIdForOther);
-
-        final String actualNotificationId1 =
-            govNotifyNotificationSender.sendEmail(
-                templateId,
-                emailAddress,
-                personalisation,
-                reference
-            );
-
-        final String actualNotificationId2 =
-            govNotifyNotificationSender.sendEmail(
-                templateId,
-                emailAddress,
-                personalisation,
-                reference
-            );
-
-        final String actualNotificationIdForOther =
-            govNotifyNotificationSender.sendEmail(
-                templateId,
-                otherEmailAddress,
-                personalisation,
-                otherReference
-            );
-
-        assertEquals(expectedNotificationId.toString(), actualNotificationId1);
-        assertEquals(expectedNotificationId.toString(), actualNotificationId2);
-        assertEquals(expectedNotificationIdForOther.toString(), actualNotificationIdForOther);
-
-        try {
-            await().atMost(2, TimeUnit.SECONDS).until(() -> false);
-        } catch (ConditionTimeoutException e) {
-            assertTrue(true, "We expect this to timeout");
-        }
-
-        final String actualNotificationId3 =
-            govNotifyNotificationSender.sendEmail(
-                templateId,
-                emailAddress,
-                personalisation,
-                reference
-            );
-
-        assertEquals(expectedNotificationId.toString(), actualNotificationId3);
-
-        verify(notificationClient, times(2)).sendEmail(
-            templateId,
-            emailAddress,
-            personalisation,
-            reference
-        );
-
-        verify(notificationClient, times(1)).sendEmail(
-            templateId,
-            otherEmailAddress,
-            personalisation,
-            otherReference
-        );
-    }
-
-    @Test
-    public void wrap_appeal_gov_notify_email_exceptions() throws NotificationClientException {
-
-        NotificationClientException underlyingException = mock(NotificationClientException.class);
-
-        doThrow(underlyingException)
-            .when(notificationClient)
-            .sendEmail(
-                templateId,
-                emailAddress,
-                personalisation,
-                reference
-            );
-
-        assertThatThrownBy(() ->
-            govNotifyNotificationSender.sendEmail(
-                templateId,
-                emailAddress,
-                personalisation,
-                reference
-            )
-        ).isExactlyInstanceOf(NotificationServiceResponseException.class)
-            .hasMessage("Failed to send email using GovNotify")
-            .hasCause(underlyingException);
-
     }
 
     @Test
@@ -202,16 +88,15 @@ public class GovNotifyNotificationSenderTest {
 
         final UUID expectedNotificationId = UUID.randomUUID();
 
-        SendSmsResponse sendSmsResponse = mock(SendSmsResponse.class);
-
-        when(notificationClient.sendSms(
-            templateId,
-            phoneNumber,
-            personalisation,
-            reference
-        )).thenReturn(sendSmsResponse);
-
-        when(sendSmsResponse.getNotificationId()).thenReturn(expectedNotificationId);
+        when(senderHelper.sendSms(
+                templateId,
+                phoneNumber,
+                personalisation,
+                reference,
+                notificationClient,
+                deduplicateSendsWithinSeconds,
+                LOG
+        )).thenReturn(String.valueOf(expectedNotificationId));
 
         String actualNotificationId =
             govNotifyNotificationSender.sendSms(
@@ -221,128 +106,17 @@ public class GovNotifyNotificationSenderTest {
                 reference
             );
 
-        verify(notificationClient, times(1)).sendSms(
+        verify(senderHelper, times(1)).sendSms(
                 templateId,
                 phoneNumber,
                 personalisation,
-                reference
+                reference,
+                notificationClient,
+                deduplicateSendsWithinSeconds,
+                LOG
         );
 
         assertEquals(expectedNotificationId.toString(), actualNotificationId);
     }
 
-    @Test
-    public void should_not_send_duplicate_sms_in_short_space_of_time() throws NotificationClientException {
-
-        final String otherPhoneNumber = "07123456780";
-        final String otherReference = "1111_SOME_OTHER_NOTIFICATION";
-
-        final UUID expectedNotificationId = UUID.randomUUID();
-        final UUID expectedNotificationIdForOther = UUID.randomUUID();
-
-        SendSmsResponse sendSmsResponse = mock(SendSmsResponse.class);
-        SendSmsResponse sendSmsResponseForOther = mock(SendSmsResponse.class);
-
-        when(notificationClient.sendSms(
-            templateId,
-            phoneNumber,
-            personalisation,
-            reference
-        )).thenReturn(sendSmsResponse);
-
-        when(notificationClient.sendSms(
-            templateId,
-            otherPhoneNumber,
-            personalisation,
-            otherReference
-        )).thenReturn(sendSmsResponseForOther);
-
-        when(sendSmsResponse.getNotificationId()).thenReturn(expectedNotificationId);
-        when(sendSmsResponseForOther.getNotificationId()).thenReturn(expectedNotificationIdForOther);
-
-        final String actualNotificationId1 =
-            govNotifyNotificationSender.sendSms(
-                templateId,
-                phoneNumber,
-                personalisation,
-                reference
-            );
-
-        final String actualNotificationId2 =
-            govNotifyNotificationSender.sendSms(
-                templateId,
-                phoneNumber,
-                personalisation,
-                reference
-            );
-
-        final String actualNotificationIdForOther =
-            govNotifyNotificationSender.sendSms(
-                templateId,
-                otherPhoneNumber,
-                personalisation,
-                otherReference
-            );
-
-        assertEquals(expectedNotificationId.toString(), actualNotificationId1);
-        assertEquals(expectedNotificationId.toString(), actualNotificationId2);
-        assertEquals(expectedNotificationIdForOther.toString(), actualNotificationIdForOther);
-
-        try {
-            await().atMost(2, TimeUnit.SECONDS).until(() -> false);
-        } catch (ConditionTimeoutException e) {
-            assertTrue(true, "We expect this to timeout");
-        }
-
-        final String actualNotificationId3 =
-            govNotifyNotificationSender.sendSms(
-                templateId,
-                phoneNumber,
-                personalisation,
-                reference
-            );
-
-        assertEquals(expectedNotificationId.toString(), actualNotificationId3);
-
-        verify(notificationClient, times(2)).sendSms(
-            templateId,
-            phoneNumber,
-            personalisation,
-            reference
-        );
-
-        verify(notificationClient, times(1)).sendSms(
-            templateId,
-            otherPhoneNumber,
-            personalisation,
-            otherReference
-        );
-    }
-
-    @Test
-    public void wrap_appeal_gov_notify_sms_exceptions() throws NotificationClientException {
-
-        NotificationClientException underlyingException = mock(NotificationClientException.class);
-
-        doThrow(underlyingException)
-            .when(notificationClient)
-            .sendSms(
-                templateId,
-                phoneNumber,
-                personalisation,
-                reference
-            );
-
-        assertThatThrownBy(() ->
-            govNotifyNotificationSender.sendSms(
-                templateId,
-                phoneNumber,
-                personalisation,
-                reference
-            )
-        ).isExactlyInstanceOf(NotificationServiceResponseException.class)
-            .hasMessage("Failed to send sms using GovNotify")
-            .hasCause(underlyingException);
-
-    }
 }
